Meti las clases en packages para que estuviese mas ordenado porque sino un quilombo saludos.
