--- MainApp.java
@@ -1,7 +1,0 @@
-import java.util.ArrayList;
-
-public class MainApp {
-    public static void main(String[] args) {
-        ArrayList<Item> items = null;
-    }
-}
