+++ entity/Entidad.java
@@ -1,10 +1,14 @@
+package entity;
+
+import main.Modificar;
+
 public abstract class Entidad extends Modificar{
     protected String nombre;
-    protected int vida;
-    protected int stamina;
-    protected float ataque;
-    protected float defensa;
-    protected float inteligencia;
+    public int vida;
+    public int stamina;
+    public float ataque;
+    public float defensa;
+    public float inteligencia;
 
     public Entidad(String nombre, int vida, int stamina, float ataque, float defensa, float inteligencia){
         this.nombre = nombre;
