+++ item/Item.java
@@ -1,3 +1,7 @@
+package item;
+
+import main.Modificar;
+
 public abstract class Item extends Modificar{
     protected String nombre;
