+++ item/acessorio/Accesorio.java
@@ -1,3 +1,7 @@
+package item.acessorio;
+
+import item.Item;
+
 public abstract class Accesorio extends Item{
 
     protected Accesorio(String nombre) {
