+++ item/arma/Arma.java
@@ -1,6 +1,10 @@
+package item.arma;
+
+import item.Item;
+
 public abstract class Arma extends Item{
-    protected float ataque;
-    protected float condicion;
+    public float ataque;
+    public float condicion;
 
     protected Arma(String nombre, float ataque, float condicion) {
         super(nombre);
