+++ main/MainApp.java
@@ -1,0 +1,10 @@
+package main;
+
+import item.Item;
+import java.util.List;
+
+public class MainApp {
+    public static void main(String[] args) {
+        List<Item> items = null;
+    }
+}
